# apps/backend/cmd/docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.Response-any:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
  controller.Response-array_service_UserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/service.UserResponse'
        type: array
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
  controller.Response-bool:
    properties:
      data:
        type: boolean
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
...
  controller.Response-service_UserResponse:
    properties:
      data:
        $ref: '#/definitions/service.UserResponse'
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
...
  service.Image:
    properties:
      created_at:
        type: string
      height:
        type: integer
      id:
        type: integer
      path:
        type: string
      size:
        type: number
      type:
        type: string
      updated_at:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
...
  service.ProfileImage:
    properties:
      created_at:
        type: string
      id:
        type: integer
      image:
        $ref: '#/definitions/service.Image'
      image_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  service.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  service.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verifyed:
        type: boolean
      full_name:
        type: string
      id:
        type: integer
      profile_image:
        items:
          $ref: '#/definitions/service.ProfileImage'
        type: array
      role:
        $ref: '#/definitions/service.Role'
      role_id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server caller server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /users/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response-service_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - users
  /users/by-username:
    get:
      consumes:
      - application/json
      parameters:
      - description: Username
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response-array_service_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response-service_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - users
  /users/profile-image:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Profile image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.Response-service_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - users
  /users/verify-email:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Response-bool'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response-any'
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"